payment-request/constructor/typeerror-on-empty-methoddata.html

<!DOCTYPE HTML>
<meta charset='utf-8'>
<title>PaymentRequest: TypeError on empty methodData.</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<script>
async_test((t) => {
  onload = t.step_func_done(() => {
    assert_throws(new TypeError(), function() {
      let pr = new PaymentRequest([], {
        id: 'foobar',
        total: {label: 'label', amount: {currency: 'USD', value: '5.00'}}
      });
    })
  });
});
</script>
